requirements.txt
pycryptodome
docopt
Flask
Flask-SocketIO
pyte